+++ thoth-export-server/src/record.rs
@@ -2,7 +2,6 @@ use actix_web::{http::StatusCode, HttpRequest, Responder};
 use csv::QuoteStyle;
 use paperclip::actix::web::HttpResponse;
 use paperclip::actix::OperationModifier;
-use paperclip::util::{ready, Ready};
 use paperclip::v2::models::{DefaultOperationRaw, Either, Response};
 use paperclip::v2::schema::Apiv2Schema;
 use std::str::FromStr;
@@ -11,7 +10,10 @@ use thoth_errors::{ThothError, ThothResult};
 
 use crate::bibtex::{BibtexSpecification, BibtexThoth};
 use crate::csv::{CsvSpecification, CsvThoth, KbartOclc};
-use crate::xml::{Onix21EbscoHost, Onix3Jstor, Onix3Oapen, Onix3ProjectMuse, XmlSpecification};
+use crate::xml::{
+    DoiDepositCrossref, Onix21EbscoHost, Onix3GoogleBooks, Onix3Jstor, Onix3Oapen,
+    Onix3ProjectMuse, XmlSpecification,
+};
 
 pub(crate) trait AsRecord {}
 impl AsRecord for Vec<Work> {}
@@ -25,10 +27,12 @@ pub(crate) enum MetadataSpecification {
     Onix3ProjectMuse(Onix3ProjectMuse),
     Onix3Oapen(Onix3Oapen),
     Onix3Jstor(Onix3Jstor),
+    Onix3GoogleBooks(Onix3GoogleBooks),
     Onix21EbscoHost(Onix21EbscoHost),
     CsvThoth(CsvThoth),
     KbartOclc(KbartOclc),
     BibtexThoth(BibtexThoth),
+    DoiDepositCrossref(DoiDepositCrossref),
 }
 
 pub(crate) struct MetadataRecord<T: AsRecord> {
@@ -63,10 +67,12 @@ where
             MetadataSpecification::Onix3ProjectMuse(_) => Self::XML_MIME_TYPE,
             MetadataSpecification::Onix3Oapen(_) => Self::XML_MIME_TYPE,
             MetadataSpecification::Onix3Jstor(_) => Self::XML_MIME_TYPE,
+            MetadataSpecification::Onix3GoogleBooks(_) => Self::XML_MIME_TYPE,
             MetadataSpecification::Onix21EbscoHost(_) => Self::XML_MIME_TYPE,
             MetadataSpecification::CsvThoth(_) => Self::CSV_MIME_TYPE,
             MetadataSpecification::KbartOclc(_) => Self::TXT_MIME_TYPE,
             MetadataSpecification::BibtexThoth(_) => Self::BIB_MIME_TYPE,
+            MetadataSpecification::DoiDepositCrossref(_) => Self::XML_MIME_TYPE,
         }
     }
 
@@ -75,10 +81,12 @@ where
             MetadataSpecification::Onix3ProjectMuse(_) => self.xml_file_name(),
             MetadataSpecification::Onix3Oapen(_) => self.xml_file_name(),
             MetadataSpecification::Onix3Jstor(_) => self.xml_file_name(),
+            MetadataSpecification::Onix3GoogleBooks(_) => self.xml_file_name(),
             MetadataSpecification::Onix21EbscoHost(_) => self.xml_file_name(),
             MetadataSpecification::CsvThoth(_) => self.csv_file_name(),
             MetadataSpecification::KbartOclc(_) => self.txt_file_name(),
             MetadataSpecification::BibtexThoth(_) => self.bib_file_name(),
+            MetadataSpecification::DoiDepositCrossref(_) => self.xml_file_name(),
         }
     }
 
@@ -124,6 +132,9 @@ impl MetadataRecord<Vec<Work>> {
             MetadataSpecification::Onix3Jstor(onix3_jstor) => {
                 onix3_jstor.generate(&self.data, None)
             }
+            MetadataSpecification::Onix3GoogleBooks(onix3_google_books) => {
+                onix3_google_books.generate(&self.data, None)
+            }
             MetadataSpecification::Onix21EbscoHost(onix21_ebsco_host) => {
                 onix21_ebsco_host.generate(&self.data, Some(DOCTYPE_ONIX21_REF))
             }
@@ -134,24 +145,23 @@ impl MetadataRecord<Vec<Work>> {
                 kbart_oclc.generate(&self.data, QuoteStyle::Necessary, DELIMITER_TAB)
             }
             MetadataSpecification::BibtexThoth(bibtex_thoth) => bibtex_thoth.generate(&self.data),
+            MetadataSpecification::DoiDepositCrossref(doideposit_crossref) => {
+                doideposit_crossref.generate(&self.data, None)
+            }
         }
     }
 }
 
-impl Responder for MetadataRecord<Vec<Work>>
-where
-    actix_web::dev::Body: From<String>,
-{
-    type Error = ThothError;
-    type Future = Ready<ThothResult<HttpResponse>>;
+impl Responder for MetadataRecord<Vec<Work>> {
+    type Body = actix_web::body::BoxBody;
 
-    fn respond_to(self, _: &HttpRequest) -> Self::Future {
+    fn respond_to(self, _: &HttpRequest) -> HttpResponse {
         match self.generate() {
-            Ok(record) => ready(Ok(HttpResponse::build(StatusCode::OK)
+            Ok(record) => HttpResponse::build(StatusCode::OK)
                 .content_type(self.content_type())
-                .header("Content-Disposition", self.content_disposition())
-                .body(record))),
-            Err(e) => ready(Err(e)),
+                .append_header(("Content-Disposition", self.content_disposition()))
+                .body(record),
+            Err(e) => HttpResponse::from_error(e),
         }
     }
 }
@@ -185,12 +195,18 @@ impl FromStr for MetadataSpecification {
             }
             "onix_3.0::oapen" => Ok(MetadataSpecification::Onix3Oapen(Onix3Oapen {})),
             "onix_3.0::jstor" => Ok(MetadataSpecification::Onix3Jstor(Onix3Jstor {})),
+            "onix_3.0::google_books" => {
+                Ok(MetadataSpecification::Onix3GoogleBooks(Onix3GoogleBooks {}))
+            }
             "onix_2.1::ebsco_host" => {
                 Ok(MetadataSpecification::Onix21EbscoHost(Onix21EbscoHost {}))
             }
             "csv::thoth" => Ok(MetadataSpecification::CsvThoth(CsvThoth {})),
             "kbart::oclc" => Ok(MetadataSpecification::KbartOclc(KbartOclc {})),
             "bibtex::thoth" => Ok(MetadataSpecification::BibtexThoth(BibtexThoth {})),
+            "doideposit::crossref" => Ok(MetadataSpecification::DoiDepositCrossref(
+                DoiDepositCrossref {},
+            )),
             _ => Err(ThothError::InvalidMetadataSpecification(input.to_string())),
         }
     }
@@ -202,10 +218,12 @@ impl ToString for MetadataSpecification {
             MetadataSpecification::Onix3ProjectMuse(_) => "onix_3.0::project_muse".to_string(),
             MetadataSpecification::Onix3Oapen(_) => "onix_3.0::oapen".to_string(),
             MetadataSpecification::Onix3Jstor(_) => "onix_3.0::jstor".to_string(),
+            MetadataSpecification::Onix3GoogleBooks(_) => "onix_3.0::google_books".to_string(),
             MetadataSpecification::Onix21EbscoHost(_) => "onix_2.1::ebsco_host".to_string(),
             MetadataSpecification::CsvThoth(_) => "csv::thoth".to_string(),
             MetadataSpecification::KbartOclc(_) => "kbart::oclc".to_string(),
             MetadataSpecification::BibtexThoth(_) => "bibtex::thoth".to_string(),
+            MetadataSpecification::DoiDepositCrossref(_) => "doideposit::crossref".to_string(),
         }
     }
 }
@@ -263,6 +281,15 @@ mod tests {
         );
         let to_test = MetadataRecord::new(
             "some_id".to_string(),
+            MetadataSpecification::Onix3GoogleBooks(Onix3GoogleBooks {}),
+            vec![],
+        );
+        assert_eq!(
+            to_test.file_name(),
+            "onix_3.0__google_books__some_id.xml".to_string()
+        );
+        let to_test = MetadataRecord::new(
+            "some_id".to_string(),
             MetadataSpecification::Onix21EbscoHost(Onix21EbscoHost {}),
             vec![],
         );
@@ -284,6 +311,15 @@ mod tests {
         assert_eq!(
             to_test.file_name(),
             "bibtex__thoth__some_id.bib".to_string()
+        );
+        let to_test = MetadataRecord::new(
+            "some_id".to_string(),
+            MetadataSpecification::DoiDepositCrossref(DoiDepositCrossref {}),
+            vec![],
+        );
+        assert_eq!(
+            to_test.file_name(),
+            "doideposit__crossref__some_id.xml".to_string()
         );
     }
 }